requirements.txt
fastapi
uvicorn
pyswisseph
stripe
